refactor move game contexts to their own folder
