+++ frontend/src/components/(game)/playmat/Board/index.tsx
@@ -1,9 +1,9 @@
 import { useDroppable } from "@dnd-kit/core";
-import { useShahrazadGameContext } from "@/contexts/game";
+import { useShahrazadGameContext } from "@/contexts/(game)/game";
 import { ShahrazadZoneId } from "@/types/bindings/zone";
 import { IDroppableData } from "@/types/interfaces/dnd";
 import { BoardCard } from "./board-card";
-import { useSelection } from "@/contexts/selection";
+import { useSelection } from "@/contexts/(game)/selection";
 import Selection from "./selection";
 
 import "./board.css";
